マージしてください。

+++ src/hooks/codes/editors/sidebar/useDeleteFileInSidebar.ts
@@ -1,24 +1,25 @@
 import { File } from "@/src/types";
-
 import { useSupabase } from "@/src/contexts/SupabaseProvider";
 import { fetchDeleteFile } from "@/src/libs/externals/supabase/queries/files";
-import { useDeleteFileFromEditorFiles } from "./useDeleteFileFromEditorFiles";
-import { useSetEditorSelectedFile } from "./setter/useSetEditorSelectedFile";
+import { useSetEditorSelectedFile } from "@/src/hooks/codes/editors/setter/useSetEditorSelectedFile";
+import { useDeleteFileFromEditorFiles } from "@/src/hooks/codes/editors/useDeleteFileFromEditorFiles";
 
-export const useDeleteFileInEditor = () => {
+export const useDeleteFileInSidebar = () => {
     const { client } = useSupabase();
 
     const { setSelectedFile } = useSetEditorSelectedFile();
     const { deleteFile } = useDeleteFileFromEditorFiles();
 
-    const deleteFileInEditor = async (file: File) => {
+
+    const onDeleteFile = async (file: File) => {
         if (!client) return;
+        if (!confirm("本当に削除しますか？")) return;
         await fetchDeleteFile(file?.id, client);
         deleteFile(file);
         setSelectedFile(undefined);
     };
 
     return {
-        deleteFileInEditor
+        onDeleteFile
     }
 }
